+++ CHANGELOG.md
@@ -4,14 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.51.0] - 2019-08-30
+### Added
+- CXV1-18685 - Message Templates component RichText Editor.
+
 ## [0.50.0] - 2019-08-29
 ### Added
 - CXV1-18753-Added-WordWrapIcon. 
+- CXV1-18685 - Message Templates component RichText Editor.
 
 ##[0.49.3] - 2019-08-28
 ### Changed
 - <no-jira> - Fixing an issue with TimepickerField and DatepickerField not updating the touched flag
-
 
 ##[0.49.2] - 2019-08-27
 ### Changed